initial commit
This web application is simply in HTML, CSS, and JavaScript. Taking about the features of this system, the user can forecast the condition of the current-day climate probability and predict whether the day is a cloudy or sunny day. The user can write down the name of any city across the world in the app. Also, this project includes a lot of javascript for making the functioning of the project.